Merge translated docs ly snippets into user/out-www
This allows ly snippets present in a translated manual but not in the
English manual to show in HTML manual.  Such ly snippets are either
the result of an outdated translation or the choice of a translator;
snippets in first case can be updated automatically (look for
snippet-update in Documentation/TRANSLATION), and the number of
snippets in second case can be estimated by greping '@c KEEP LY' in
translated .itely's.