Convert dynamic marks (given in a <direction> tag, assigned to the staff at a given position in xml, not to a note like in lilypond)
In the LilyPondVoiceBuilder, I added a method to store any pending dynamics and print them out only after the next note or rest (everything with duration>0) is encountered.

Also convert (de-)crescendo (begin/end also given in a <direction> tag, not assigned to a particular note)

Comment about broken dynamics, when they appear as first element of a part before any note (so that no voice_id is known yet).