Fix translation infrastructure
The fixes in html-gettext are very ugly, but I'm sick and tired of
fuzzing with regular expressions for cheating with node names.  I'm
not sure all this dirty trickery works, I'm going to contact Texinfo
developers for feature requests, so we can get rid of this mess of
Python and Gettext.