MusicXML: Fine-tune chord detection in combination with dynamics
When trying to find the corresponding notes for notes with the <chord>
indicator, we need to skip all dynamics, but must not cross bar lines. In
particular, since we now also add dynamics events, we cannot automatically
take the last event, but have to look for the last EventChord.

Also some small bug fixes (need to quote texts with _, missing param in dead
code, style, etc.)

Signed-off-by: Reinhold Kainhofer <reinhold@kainhofer.com>