MusicXML: Allow .xml to be left out on input files, allow .ly on output files
The output file name is assumed to be without the .ly suffix, which is
counterintuitive: As as user I would expect the option "-o test.ly" to
generate a file test.ly. However, .ly was always appended to the filename,
so remove a possible .ly suffix if given on the output filename.

Similarly, the input file name was assumed to be the full file name. In
particular with tab-completion in a shell, this means several more keystrokes.
If the input file given does not exists, I now also check if a file
exists with the .xml suffix appended. This is similar to the behavior of
lilypond, where "lilypond test" will also convert a file test.ly.

Signed-off-by: Reinhold Kainhofer <reinhold@kainhofer.com>