MusicXML: Fix graces to allow beams, chords, etc.
Instead of keeping an array of grace elements for each chord,
rather use a full SequentialMusic object that holds EventChords
(so we can insert beams, real grace chords, etc.). For this we also
need a method in the SequentialMusic class to get the last chord.
Depending on the slash attribute of the <grace> element, we now
use acciaccatura rather than grace for slashed grace notes.

Signed-off-by: Reinhold Kainhofer <reinhold@kainhofer.com>