MusicXML: Add support for compressed MusicXML files (command-line option -z)
MusicXML also defined a compressed format, where the .xml file is
stored inside a ZIP archive (with extension .mxl) . In this case, we need to
read the META-INF/container.xml file from the archive to get the name
of the XML file inside the archive, and then load this file. Thanks
to the zipfile module and the StringIO module, implementing this was quite easy.

I also fixed some other small bugs that led to crashes.