Partcombine: Fix crash when empty music is passed
I usually start with the general structure of a large score
and only then fill it with music. This means that partcombine
is fed two empty music expressions (or only containing one
multi-measure rest each), so that evs1 and evs2 won't contain
data and (assoc "one" evs1) will be an empty list => cdr on
it will crash.
In this case, simply set the split list to an empty list and
don't crash...

Signed-off-by: Reinhold Kainhofer <reinhold@kainhofer.com>