MusicXML: Add basic support for figured bass
The figured bass definitionis correctly generated and written to the -defs.ly
file, but not yet integrated into the score itself.

For this, I'm keeping a second voice builder while going through the voices'
children and insert the figured bass there and at the end extract the
definition from it.