figured bass: Implement backslashed figures (raised 6th steps)
raised 6th steps are typically shown as 6 with a backslash through it
(rather than with a '+'). This patch implements this by adding the
\\ modifier to bass figures: <6\\>

The back-slashed digit is printed out using the new markup function
backslashed-digit (copied from slashed-digit and adapted to backslashes).
Both forward and backward slashes uses the same internal function.

I also increased the slope of forward slashes to make it easier to
distinguish the direction of the slash.

For some numbers I also enlarge the slash in X directions to make it
more visible (my hand-engraved scores use an even larger slash than
lilypond, anyway). I also move the slash up/down for some numbers
to make it more visible.

Also use the real Y extents of the slash rather than the whole interval
for the number. As we shift the slash for some numbers, otherwise
it would make the whole digit larger and shift it down for no apparent
visual reason.

Signed-off-by: Reinhold Kainhofer <reinhold@kainhofer.com>