texi2html: Remove code to create _xref.html; Load and use the _xref.html files
Removed the code to generate the *_xref.html file, because that's now done by
the extract_texi_filenames.py script before texi2html is run.
However, I implemented loading those files (both for the nodes of the
current files and for xrefs to other files). Somehow I can't make Perl
find and load the file, so it does not work yet... Finding the files for
xrefs is not yet even implemented...

I could probably remove the code to determine the file name in the
texi2html init file, but I decided to leave it in there, just in
case something went wrong with the *_xref.map files. In that case,
we'll still get correct filenames/anchors, but of course xrefs to
those sections from other manuals will not be able to determine the
filename/anchor...