Feature: Add Harp pedal diagrams, i.e. a \harp-pedal markup function
Add a \harp-pedal markup function, which produces harp pedalling diagrams. The
syntax is
    \harp-pedal #"^-v|--ov^"
The contents of the string define the positions of the pedals (^/-/v for
up/neutral/down), | is the divider and o indicates that the next pedal
should be circled. It can be tweaked through the size and thickness props
and the harp-pedal-details property list of the TextScript grob. For this, I
had to add a harp-pedal-interface, too.

The only issue I'm still having is that lilypond does not have a function
to draw an ellipse, so all I can do for circled pedals is to draw a circle
rather than an ellipse... Also, using a rounded box does not work, as that
fills its contents, thus overdrawing the pedal box.

I've also added three regression test cases, namely
-) Basic functionality checks (including circled boxes and an empty definition
   string, as well as invalid characters in the string)
-) Sanity checks: Lilypond will check whether the diagram follows standard
   rules (7 pedals with one divider after the third). If not, it will print
   out a warning, but should still produce the diagram as given.
-) Tweaking: Most details of the diagram can be tweaked using size, thickness
   and harp-pedal-details.