Fix a ragged-bottom/ragged-last-bottom corner case.
When using ragged spacing, we want to pretend that the last page
requires no force (so that we don't worsen the spacing of the
previous pages in order to fill the last page). The previous check
for this only worked if ragged-last-bottom = ##t, but not if
ragged-bottom = ##t and ragged-last-bottom ##f.