Introduce @endpredefined.
Fix @funindex as a preparation for using texi2html to expand macros.
Note that both makeinfo and texi2dvi are buggy currently and can't do
what we need :-(  We have to wait until texi2html is in shape so that
it can replace `makeinfo -E ...'.  Right now, it has some syntax issues
which prevents it, but the maintainer is already going to fix them.

Formatting issues; adding some comments.