use correct left to right ordering in keySignature list
The keySignature property list of the KeySignature so far was using
reversed order (e.g. the rightmost accidental came first). This was
quite confusing when manually setting non-standard key signatures like
  \set Staff.keySignature = #`(((0 . 3) . ,SHARP)
                               ((0 . 5) . ,FLAT)
                               ((0 . 6) . ,FLAT))
In this example, you got bes, as, fis in that reversed order rather than
in the order in which you list the alterations. Now, you'll get fis, as,
bes in that (expected) order.

Also update the snippets and add convert-ly check to warn users about the
reversion (the keySignature property was hardly ever used, though).