Change the treatment of min-distance in springs.
Previously, we ensured that the ideal distance was at least
as large as the minimum distance, but it sometimes useful
not to. Now, if a spring has minimum distance larger than
ideal distance, it will remain fixed (at minimum-distance)
for a while as you start to stretch it, but it will stretch
eventually. This helps give us a more uniform vertical
spacing.