Adjust "font-name-style" procedure for SVG backend
Changes to make this procedure more generic for matching both fonts
with string names and font smobs that need their names extracted.

In the future, if all fonts are identified with string names, and
the "feta-alphabet" PFB fonts are no longer used, this procedure may
need to be adjusted.

Note that this procedure is only used for the SVG backend.