Rewrite the vertical layout of staves/systems.
This combines the positioning of systems and staves into
a single pass of a rod-and-spring problem. As a consequence,
the spacing of staves within a system is much more configurable.

Most properties related to vertical spacing have been removed
(eg. page-top-space, between-system-space, between-system-padding)
and replaced with more powerful versions (eg. first-system-spacing,
first-system-title-spacing, between-system-spacing).