Implement framework for post-fix text (de)cresc spanners (backend only)
The main problem why \decr or \dim were prefix operators was that the
new dynamic engraver was reading only the properties of the Engraver
itself, so there was no way to create a CrescendoEvent and at the same
time specify the hairpin/text behavior.

This patch makes the New_dynamic_engraver also look at the properties
of the (Dec|C)recendoEvent and prefer a setting from there. If the event
does not have a property (which will be the case most of the time), the
Engraver's settings are used (i.e. same behavior as now).

With this change, e.g. \dim can be defined as
dim = #(make-music 'DecrescendoEvent 'span-direction START
            'span-type text 'span-text "dim.")
and used as postfix like a4\dim without the need of any overrides
before the affected note.

This is just the backend commit, implementing processing of the event's
properties. The definition of \cresc is still unchanged, leaving the old
prefix notation in place, until we find  good upgrade path.

Also add regtests for text cresc spanners and add the regtest to the
snippets.