Use the full pure Y-extent for complete systems.
For staves within a system, we split up the pure Y-extent into the Y-extent of
the prefatory matter and the Y-extent of the rest (thereby emulating a crude
skyline).  Between systems, we use a single extent box.  Before this commit, we
were ignoring the Y-extent of the prefatory matter altogether; now the extent
box is correct.