Fix `split-at-predicate' in `scm/lily-library.scm'.
Don't make the assumption that if `(PRED previous_element element)'
is false, `(PRED element previous_element)' must be true.

Suppose we want to split `(1 2 3 5 7)' into two lists such that
the first list contains the initial consecutive integers `(1 2 3)'
and the second contains the remaining elements `(5 7)'.  We expect
`(split-at-predicate (lambda (x y) (= (- y x) 1)) '(1 2 3 5 7))'
to return `((1 2 3) . (5 7))', but in fact it returns `((1 2 3 5 7))'
because of this erroneous assumption.