Pango: Do not assume markup is only LTR or RTL.
The current code assumes that markup is either LTR or RTL.  But it could
be both.  For example

  \markup { "h א" }

should be displayed as written, but LilyPond displays this string as

  "אh"

This patch fixes that issue.

With pango_itemize(), Pango automatically (and correctly) divides the
markup into sections that are intended to be processed by the Unicode
Bidirectional Algorithm.  Pango also handles the RTL character
reordering, within each of these sections, if necessary.

For now, until we implement the Unicode Bidirectional Algorithm, or
another solution is found, simply place each section in order from left
to right.