Web build: separate makefile for security + ease.
We can't rely on the normal makefile stuff, since that 1) involves
a lot of extra building on lilypond.org, and 2) it's a security
risk.  So anything that needs to be done for the automatic hourly
builds should be separated as much as possible, since I need to
manually look through any files used to check for security flaws.
I also need to do this whenever those files are updated.