Applied the patch from Neil Puttock to tag  release/2.11.45-1:
0001-Draw-bar-lines-correctly-when-staff-is-not-centered.patch

which was posted to lilypond-dev on 7/4/08:

Hi Victor,

2008/7/2 V!ctor Adán <vctradn@gmail.com>:
> Hello All,
>
> I'm using Staff contexts with the "line-positions" property modified to fit
> my capricious needs.
> In doing so I'm running into 2 problems (please see attached png):
>
> 1. When I set lines to odd numbers, the duration dots of the note heads fall
> *on* the staff lines rather than within the spaces. It's like Lilypond is
> hard-coded to put the dots on the even number slots rather that looking at
> the "line-positions" attribute to decide on the best place to put these
> dots. Is there a way to change this?
>
> 2. It seems that the line-positions must be centered at 0, otherwise (as in
> the example below), the bar line is drawn off-center, either moved up or
> down. Is there a way to have the barline fit the boundaries of the two
> outermost staff lines exactly without having the line-positions 0-centered?
> I found this mail by Kevin Dalley in the mailing list, so it seems this has
> been an issue before...
> http://www.nabble.com/Does-the-center-of-the-staff-need-to-be-zero--p9507190.html

It's a good thing you've brought this up, since that post from Kevin
and another he posted a few days later include patches which fix both
of the issues you're experiencing; there were some problems with the
patches which prevented them being applied for testing, so they seem
to have fallen by the wayside.

I certainly think they would be a useful improvement on the current
behaviour, so I've tidied the patches up to make them work with the
current version; attached is the output of your snippet with the
patches applied.

Regards,
Neil